Working on better managment of empty files
Currently the editor winds up adding a 0, or making an empty line have
at least one \n.  The behavior should be to leave a file untouched, and
be happy with a file that only contains characters and no line end.

So this is that cleanup, since it isn't as simple as it
sounds due to assumuptions I made about what constitutes a valid buffer
and a selection of lines.  Now the editor is more correct about handling
ranges and line ends.

Also added a new test for Append-A on empty line, and fix for it